+++ src/openassetio-core/include/openassetio/errors/exceptions.hpp
@@ -95,16 +95,15 @@ struct OPENASSETIO_CORE_EXPORT UnhandledException : OpenAssetIOException {
  */
 
 /**
- * Exception base that ties together a @ref errors.BatchElementError and
- * an index.
+ * Exception thrown when a @ref BatchElementError is converted.
  *
- * When thrown from a function, indicates that a particular
- * element has caused an error. The specific element that has errored
- * is indicated by the index attribute, relative to the input container.
+ * Not a type that a manager should throw, exclusively thrown via the
+ * middleware when the user is calling an exceptional convenience and a
+ * @ref BatchElementError is emitted by the manager.
  */
-struct OPENASSETIO_CORE_EXPORT BatchElementException : std::runtime_error {
-  BatchElementException(std::size_t idx, BatchElementError err)
-      : std::runtime_error{err.message}, index{idx}, error{std::move(err)} {}
+struct OPENASSETIO_CORE_EXPORT BatchElementException : OpenAssetIOException {
+  BatchElementException(std::size_t idx, BatchElementError err, const std::string& message)
+      : OpenAssetIOException{message}, index{idx}, error{std::move(err)} {}
 
   /**
    * Index describing which batch element has caused an error.
@@ -115,64 +114,6 @@ struct OPENASSETIO_CORE_EXPORT BatchElementException : std::runtime_error {
    * Object describing the nature of the specific error.
    */
   BatchElementError error;
-};
-
-/**
- * Exception equivalent of
- * @ref errors.BatchElementError.ErrorCode.kUnknown
- */
-struct OPENASSETIO_CORE_EXPORT UnknownBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref errors.BatchElementError.ErrorCode.kInvalidEntityReference
- */
-struct OPENASSETIO_CORE_EXPORT InvalidEntityReferenceBatchElementException
-    : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref errors.BatchElementError.ErrorCode.kMalformedEntityReference
- */
-struct OPENASSETIO_CORE_EXPORT MalformedEntityReferenceBatchElementException
-    : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref errors.BatchElementError.ErrorCode.kEntityAccessError
- */
-struct OPENASSETIO_CORE_EXPORT EntityAccessErrorBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref errors.BatchElementError.ErrorCode.kEntityResolutionError
- */
-struct OPENASSETIO_CORE_EXPORT EntityResolutionErrorBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref errors.BatchElementError.ErrorCode.kInvalidPreflightHint
- */
-struct OPENASSETIO_CORE_EXPORT InvalidPreflightHintBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref errors.BatchElementError.ErrorCode.kInvalidTraitSet
- */
-struct OPENASSETIO_CORE_EXPORT InvalidTraitSetBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
 };
 
 /**