+++ src/openassetio-core/src/hostApi/Manager.cpp
@@ -16,36 +16,11 @@
 #include <openassetio/managerApi/ManagerInterface.hpp>
 #include <openassetio/typedefs.hpp>
 
+#include "../errors/exceptionMessages.hpp"
+
 namespace openassetio {
 inline namespace OPENASSETIO_CORE_ABI_VERSION {
 namespace {
-// Takes a BatchElementError and throws an equivalent exception.
-void throwFromBatchElementError(std::size_t index, errors::BatchElementError error) {
-  switch (error.code) {
-    case errors::BatchElementError::ErrorCode::kUnknown:
-      throw errors::UnknownBatchElementException(index, std::move(error));
-    case errors::BatchElementError::ErrorCode::kInvalidEntityReference:
-      throw errors::InvalidEntityReferenceBatchElementException(index, std::move(error));
-    case errors::BatchElementError::ErrorCode::kMalformedEntityReference:
-      throw errors::MalformedEntityReferenceBatchElementException(index, std::move(error));
-    case errors::BatchElementError::ErrorCode::kEntityAccessError:
-      throw errors::EntityAccessErrorBatchElementException(index, std::move(error));
-    case errors::BatchElementError::ErrorCode::kEntityResolutionError:
-      throw errors::EntityResolutionErrorBatchElementException(index, std::move(error));
-    case errors::BatchElementError::ErrorCode::kInvalidPreflightHint:
-      throw errors::InvalidPreflightHintBatchElementException(index, std::move(error));
-    case errors::BatchElementError::ErrorCode::kInvalidTraitSet:
-      throw errors::InvalidTraitSetBatchElementException(index, std::move(error));
-    default:
-      std::string exceptionMessage = "Invalid BatchElementError. Code: ";
-      exceptionMessage += std::to_string(static_cast<int>(error.code));
-      exceptionMessage += " Message: ";
-      exceptionMessage += error.message;
-      error.message = std::move(exceptionMessage);
-      throw errors::UnknownBatchElementException{index, std::move(error)};
-  }
-}
-
 /**
  * Extract the entity reference prefix from a manager plugin's info
  * dictionary, if available.
@@ -197,8 +172,10 @@ TraitsDataPtr hostApi::Manager::resolve(
       [&resolveResult]([[maybe_unused]] std::size_t index, TraitsDataPtr data) {
         resolveResult = std::move(data);
       },
-      [](std::size_t index, errors::BatchElementError error) {
-        throwFromBatchElementError(index, std::move(error));
+      [&entityReference, resolveAccess](std::size_t index, errors::BatchElementError error) {
+        auto msg = errors::createBatchElementExceptionMessage(
+            error, index, entityReference, static_cast<access::Access>(resolveAccess));
+        throw errors::BatchElementException(index, std::move(error), msg);
       });
 
   return resolveResult;
@@ -235,9 +212,11 @@ std::vector<TraitsDataPtr> hostApi::Manager::resolve(
       [&resolveResult](std::size_t index, TraitsDataPtr data) {
         resolveResult[index] = std::move(data);
       },
-      [](std::size_t index, errors::BatchElementError error) {
+      [&entityReferences, resolveAccess](std::size_t index, errors::BatchElementError error) {
         // Implemented as if FAILFAST is true.
-        throwFromBatchElementError(index, std::move(error));
+        auto msg = errors::createBatchElementExceptionMessage(
+            error, index, entityReferences[index], static_cast<access::Access>(resolveAccess));
+        throw errors::BatchElementException(index, std::move(error), msg);
       });
 
   return resolveResult;
@@ -377,8 +356,10 @@ EntityReference Manager::preflight(
       [&result]([[maybe_unused]] std::size_t index, EntityReference preflightedRef) {
         result = std::move(preflightedRef);
       },
-      [](std::size_t index, errors::BatchElementError error) {
-        throwFromBatchElementError(index, std::move(error));
+      [&entityReference, publishingAccess](std::size_t index, errors::BatchElementError error) {
+        auto msg = errors::createBatchElementExceptionMessage(
+            error, index, entityReference, static_cast<access::Access>(publishingAccess));
+        throw errors::BatchElementException(index, std::move(error), msg);
       });
 
   return result;
@@ -413,9 +394,11 @@ EntityReferences Manager::preflight(
       [&results](std::size_t index, EntityReference preflightedRef) {
         results[index] = std::move(preflightedRef);
       },
-      [](std::size_t index, errors::BatchElementError error) {
+      [&entityReferences, publishingAccess](std::size_t index, errors::BatchElementError error) {
         // Implemented as if FAILFAST is true.
-        throwFromBatchElementError(index, std::move(error));
+        auto msg = errors::createBatchElementExceptionMessage(
+            error, index, entityReferences[index], static_cast<access::Access>(publishingAccess));
+        throw errors::BatchElementException(index, std::move(error), msg);
       });
 
   return results;
@@ -468,8 +451,10 @@ EntityReference hostApi::Manager::register_(
       [&result]([[maybe_unused]] std::size_t index, EntityReference registeredRef) {
         result = std::move(registeredRef);
       },
-      [](std::size_t index, errors::BatchElementError error) {
-        throwFromBatchElementError(index, std::move(error));
+      [&entityReference, publishingAccess](std::size_t index, errors::BatchElementError error) {
+        auto msg = errors::createBatchElementExceptionMessage(
+            error, index, entityReference, static_cast<access::Access>(publishingAccess));
+        throw errors::BatchElementException(index, std::move(error), msg);
       });
 
   return result;
@@ -506,9 +491,11 @@ std::vector<EntityReference> hostApi::Manager::register_(
       [&result](std::size_t index, EntityReference registeredRef) {
         result[index] = std::move(registeredRef);
       },
-      [](std::size_t index, errors::BatchElementError error) {
+      [&entityReferences, publishingAccess](std::size_t index, errors::BatchElementError error) {
         // Implemented as if FAILFAST is true.
-        throwFromBatchElementError(index, std::move(error));
+        auto msg = errors::createBatchElementExceptionMessage(
+            error, index, entityReferences[index], static_cast<access::Access>(publishingAccess));
+        throw errors::BatchElementException(index, std::move(error), msg);
       });
 
   return result;