+++ src/openassetio-core/tests/hostApi/ManagerTest.cpp
@@ -247,9 +247,15 @@ SCENARIO("Resolving entities") {
 
       WHEN("singular resolve is called with default errorPolicyTag") {
         THEN("an exception is thrown") {
-          CHECK_THROWS_MATCHES(manager->resolve(ref, traits, resolveAccess, context),
-                               openassetio::errors::MalformedEntityReferenceBatchElementException,
-                               Catch::Message("Error Message"));
+          CHECK_THROWS_MATCHES(
+              manager->resolve(ref, traits, resolveAccess, context),
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expected.code &&
+                           std::string(e.what()).find(expected.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("singular resolve is called with kException errorPolicyTag") {
@@ -257,8 +263,13 @@ SCENARIO("Resolving entities") {
           CHECK_THROWS_MATCHES(
               manager->resolve(ref, traits, resolveAccess, context,
                                hostApi::Manager::BatchElementErrorPolicyTag::kException),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Error Message"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expected.code &&
+                           std::string(e.what()).find(expected.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("singular resolve is called with kVariant errorPolicyTag") {
@@ -297,9 +308,15 @@ SCENARIO("Resolving entities") {
 
       WHEN("batch resolve is called with default errorPolicyTag") {
         THEN("an exception is thrown") {
-          CHECK_THROWS_MATCHES(manager->resolve(refs, traits, resolveAccess, context),
-                               openassetio::errors::MalformedEntityReferenceBatchElementException,
-                               Catch::Message("Malformed Mock Error🤖"));
+          CHECK_THROWS_MATCHES(
+              manager->resolve(refs, traits, resolveAccess, context),
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expectedError0.code &&
+                           std::string(e.what()).find(expectedError0.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("batch resolve is called with kException errorPolicyTag") {
@@ -307,8 +324,13 @@ SCENARIO("Resolving entities") {
           CHECK_THROWS_MATCHES(
               manager->resolve(refs, traits, resolveAccess, context,
                                hostApi::Manager::BatchElementErrorPolicyTag::kException),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Malformed Mock Error🤖"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expectedError0.code &&
+                           std::string(e.what()).find(expectedError0.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("batch resolve is called with kVariant errorPolicyTag") {
@@ -331,103 +353,6 @@ SCENARIO("Resolving entities") {
 }
 
 using ErrorCode = openassetio::errors::BatchElementError::ErrorCode;
-
-template <class T, ErrorCode C>
-struct BatchElementErrorMapping {
-  using ExceptionType = T;
-  static constexpr ErrorCode kErrorCode = C;
-};
-
-TEMPLATE_TEST_CASE(
-    "BatchElementError conversion to exceptions when resolving", "",
-    (BatchElementErrorMapping<openassetio::errors::UnknownBatchElementException,
-                              ErrorCode::kUnknown>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidEntityReferenceBatchElementException,
-                              ErrorCode::kInvalidEntityReference>),
-    (BatchElementErrorMapping<openassetio::errors::MalformedEntityReferenceBatchElementException,
-                              ErrorCode::kMalformedEntityReference>),
-    (BatchElementErrorMapping<openassetio::errors::EntityAccessErrorBatchElementException,
-                              ErrorCode::kEntityAccessError>),
-    (BatchElementErrorMapping<openassetio::errors::EntityResolutionErrorBatchElementException,
-                              ErrorCode::kEntityResolutionError>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidPreflightHintBatchElementException,
-                              ErrorCode::kInvalidPreflightHint>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidTraitSetBatchElementException,
-                              ErrorCode::kInvalidTraitSet>)) {
-  namespace hostApi = openassetio::hostApi;
-  using trompeloeil::_;
-
-  using ExpectedExceptionType = typename TestType::ExceptionType;
-  static constexpr ErrorCode kExpectedErrorCode = TestType::kErrorCode;
-
-  GIVEN("a configured Manager instance") {
-    const openassetio::trait::TraitSet traits = {"fakeTrait", "secondFakeTrait"};
-    const openassetio::ManagerFixture fixture;
-    const auto& manager = fixture.manager;
-    auto& mockManagerInterface = fixture.mockManagerInterface;
-    const auto& context = fixture.context;
-    const auto& hostSession = fixture.hostSession;
-
-    AND_GIVEN(
-        "manager plugin will encounter an entity-specific error when next resolving a reference") {
-      const openassetio::EntityReference ref = openassetio::EntityReference{"testReference"};
-      const openassetio::EntityReferences refs = {ref};
-      const auto resolveAccess = openassetio::access::ResolveAccess::kRead;
-
-      const openassetio::errors::BatchElementError expectedError{kExpectedErrorCode,
-                                                                 "Some error message"};
-
-      // With error callback side effect
-      REQUIRE_CALL(mockManagerInterface,
-                   resolve(refs, traits, resolveAccess, context, hostSession, _, _))
-          .LR_SIDE_EFFECT(_7(123, expectedError));
-
-      WHEN("resolve is called with kException errorPolicyTag") {
-        THEN("an exception is thrown") {
-          try {
-            manager->resolve(ref, traits, resolveAccess, context,
-                             hostApi::Manager::BatchElementErrorPolicyTag::kException);
-            FAIL_CHECK("Exception not thrown");
-          } catch (const ExpectedExceptionType& exc) {
-            CHECK(exc.what() == expectedError.message);
-            CHECK(exc.error == expectedError);
-            CHECK(exc.index == 123);
-          }
-        }
-      }
-    }
-
-    AND_GIVEN(
-        "manager plugin will encounter entity-specific errors when next resolving multiple "
-        "references") {
-      const openassetio::EntityReferences refs = {openassetio::EntityReference{"testReference1"},
-                                                  openassetio::EntityReference{"testReference2"}};
-      const auto resolveAccess = openassetio::access::ResolveAccess::kRead;
-
-      const openassetio::errors::BatchElementError expectedError{kExpectedErrorCode, "Some error"};
-
-      // With error callback side effect
-      REQUIRE_CALL(mockManagerInterface,
-                   resolve(refs, traits, resolveAccess, context, hostSession, _, _))
-          .LR_SIDE_EFFECT(_7(123, expectedError))
-          .LR_SIDE_EFFECT(FAIL_CHECK("Exception should have short-circuited this"));
-
-      WHEN("resolve is called with kException errorPolicyTag") {
-        THEN("an exception is thrown") {
-          try {
-            manager->resolve(refs, traits, resolveAccess, context,
-                             hostApi::Manager::BatchElementErrorPolicyTag::kException);
-            FAIL_CHECK("Exception not thrown");
-          } catch (const ExpectedExceptionType& exc) {
-            CHECK(exc.what() == expectedError.message);
-            CHECK(exc.error == expectedError);
-            CHECK(exc.index == 123);
-          }
-        }
-      }
-    }
-  }
-}
 
 SCENARIO("Preflighting entities") {
   namespace hostApi = openassetio::hostApi;
@@ -584,9 +509,15 @@ SCENARIO("Preflighting entities") {
 
       WHEN("singular preflight is called with default errorPolicyTag") {
         THEN("an exception is thrown") {
-          CHECK_THROWS_MATCHES(manager->preflight(ref, traitsData, publishingAccess, context),
-                               openassetio::errors::MalformedEntityReferenceBatchElementException,
-                               Catch::Message("Error Message"));
+          CHECK_THROWS_MATCHES(
+              manager->preflight(ref, traitsData, publishingAccess, context),
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expected.code &&
+                           std::string(e.what()).find(expected.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("singular preflight is called with kException errorPolicyTag") {
@@ -594,8 +525,13 @@ SCENARIO("Preflighting entities") {
           CHECK_THROWS_MATCHES(
               manager->preflight(ref, traitsData, publishingAccess, context,
                                  hostApi::Manager::BatchElementErrorPolicyTag::kException),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Error Message"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expected.code &&
+                           std::string(e.what()).find(expected.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("singular preflight is called with kVariant errorPolicyTag") {
@@ -631,8 +567,13 @@ SCENARIO("Preflighting entities") {
         THEN("an exception is thrown") {
           CHECK_THROWS_MATCHES(
               manager->preflight(threeRefs, threeTraitsDatas, publishingAccess, context),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Malformed Mock Error🤖"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expectedError0.code &&
+                           std::string(e.what()).find(expectedError0.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("batch preflight is called with kException errorPolicyTag") {
@@ -640,8 +581,13 @@ SCENARIO("Preflighting entities") {
           CHECK_THROWS_MATCHES(
               manager->preflight(threeRefs, threeTraitsDatas, publishingAccess, context,
                                  hostApi::Manager::BatchElementErrorPolicyTag::kException),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Malformed Mock Error🤖"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expectedError0.code &&
+                           std::string(e.what()).find(expectedError0.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("batch preflight is called with kVariant errorPolicyTag") {
@@ -657,101 +603,6 @@ SCENARIO("Preflighting entities") {
           CHECK(error1 == expectedError1);
 
           CHECK(std::get<openassetio::EntityReference>(actualVec[2]) == expectedValue2);
-        }
-      }
-    }
-  }
-}
-
-TEMPLATE_TEST_CASE(
-    "BatchElementError conversion to exceptions when preflighting", "",
-    (BatchElementErrorMapping<openassetio::errors::UnknownBatchElementException,
-                              ErrorCode::kUnknown>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidEntityReferenceBatchElementException,
-                              ErrorCode::kInvalidEntityReference>),
-    (BatchElementErrorMapping<openassetio::errors::MalformedEntityReferenceBatchElementException,
-                              ErrorCode::kMalformedEntityReference>),
-    (BatchElementErrorMapping<openassetio::errors::EntityAccessErrorBatchElementException,
-                              ErrorCode::kEntityAccessError>),
-    (BatchElementErrorMapping<openassetio::errors::EntityResolutionErrorBatchElementException,
-                              ErrorCode::kEntityResolutionError>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidPreflightHintBatchElementException,
-                              ErrorCode::kInvalidPreflightHint>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidTraitSetBatchElementException,
-                              ErrorCode::kInvalidTraitSet>)) {
-  namespace hostApi = openassetio::hostApi;
-  using trompeloeil::_;
-
-  using ExpectedExceptionType = typename TestType::ExceptionType;
-  static constexpr ErrorCode kExpectedErrorCode = TestType::kErrorCode;
-
-  GIVEN("a configured Manager instance") {
-    const openassetio::EntityReference ref = openassetio::EntityReference{"testReference"};
-    const openassetio::EntityReferences threeRefs = {
-        openassetio::EntityReference{"testReference1"},
-        openassetio::EntityReference{"testReference2"},
-        openassetio::EntityReference{"testReference3"}};
-
-    const openassetio::TraitsDataPtr traitsData =
-        openassetio::TraitsData::make({"fakeTrait", "secondFakeTrait"});
-    const openassetio::trait::TraitsDatas threeTraitsDatas{traitsData, traitsData, traitsData};
-
-    const openassetio::ManagerFixture fixture;
-    const auto& manager = fixture.manager;
-    auto& mockManagerInterface = fixture.mockManagerInterface;
-    const auto& context = fixture.context;
-    const auto& hostSession = fixture.hostSession;
-    const auto publishingAccess = openassetio::access::PublishingAccess::kWrite;
-
-    AND_GIVEN(
-        "manager plugin will encounter an entity-specific error when next preflighting a "
-        "reference") {
-      const openassetio::errors::BatchElementError expectedError{kExpectedErrorCode,
-                                                                 "Some error message"};
-
-      // With error callback side effect
-      REQUIRE_CALL(mockManagerInterface, preflight(openassetio::EntityReferences{ref},
-                                                   openassetio::trait::TraitsDatas{traitsData},
-                                                   publishingAccess, context, hostSession, _, _))
-          .LR_SIDE_EFFECT(_7(123, expectedError));
-
-      WHEN("preflight is called with kException errorPolicyTag") {
-        THEN("an exception is thrown") {
-          try {
-            manager->preflight(ref, traitsData, publishingAccess, context,
-                               hostApi::Manager::BatchElementErrorPolicyTag::kException);
-            FAIL_CHECK("Exception not thrown");
-          } catch (const ExpectedExceptionType& exc) {
-            CHECK(exc.what() == expectedError.message);
-            CHECK(exc.error == expectedError);
-            CHECK(exc.index == 123);
-          }
-        }
-      }
-    }
-
-    AND_GIVEN(
-        "manager plugin will encounter entity-specific errors when next preflighting multiple "
-        "references") {
-      const openassetio::errors::BatchElementError expectedError{kExpectedErrorCode, "Some error"};
-
-      // With error callback side effect
-      REQUIRE_CALL(mockManagerInterface, preflight(threeRefs, threeTraitsDatas, publishingAccess,
-                                                   context, hostSession, _, _))
-          .LR_SIDE_EFFECT(_7(123, expectedError))
-          .LR_SIDE_EFFECT(FAIL_CHECK("Exception should have short-circuited this"));
-
-      WHEN("preflight is called with kException errorPolicyTag") {
-        THEN("an exception is thrown") {
-          try {
-            manager->preflight(threeRefs, threeTraitsDatas, publishingAccess, context,
-                               hostApi::Manager::BatchElementErrorPolicyTag::kException);
-            FAIL_CHECK("Exception not thrown");
-          } catch (const ExpectedExceptionType& exc) {
-            CHECK(exc.what() == expectedError.message);
-            CHECK(exc.error == expectedError);
-            CHECK(exc.index == 123);
-          }
         }
       }
     }
@@ -927,8 +778,13 @@ SCENARIO("Registering entities") {
         THEN("an exception is thrown") {
           CHECK_THROWS_MATCHES(
               manager->register_(ref, singleTraitsData, publishingAccess, context),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Error Message"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expected.code &&
+                           std::string(e.what()).find(expected.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("singular register is called with kException errorPolicyTag") {
@@ -936,8 +792,13 @@ SCENARIO("Registering entities") {
           CHECK_THROWS_MATCHES(
               manager->register_(ref, singleTraitsData, publishingAccess, context,
                                  hostApi::Manager::BatchElementErrorPolicyTag::kException),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Error Message"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expected.code &&
+                           std::string(e.what()).find(expected.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("singular register is called with kVariant errorPolicyTag") {
@@ -977,8 +838,13 @@ SCENARIO("Registering entities") {
         THEN("an exception is thrown") {
           CHECK_THROWS_MATCHES(
               manager->register_(refs, threeTraitsDatas, publishingAccess, context),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Malformed Mock Error🤖"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expectedError0.code &&
+                           std::string(e.what()).find(expectedError0.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("batch register is called with kException errorPolicyTag") {
@@ -986,8 +852,13 @@ SCENARIO("Registering entities") {
           CHECK_THROWS_MATCHES(
               manager->register_(refs, threeTraitsDatas, publishingAccess, context,
                                  hostApi::Manager::BatchElementErrorPolicyTag::kException),
-              openassetio::errors::MalformedEntityReferenceBatchElementException,
-              Catch::Message("Malformed Mock Error🤖"));
+              openassetio::errors::BatchElementException,
+              Catch::Predicate<openassetio::errors::BatchElementException>(
+                  [&](const openassetio::errors::BatchElementException& e) -> bool {
+                    return e.error.code == expectedError0.code &&
+                           std::string(e.what()).find(expectedError0.message) != std::string::npos;
+                  },
+                  "Thrown exception has unexpected message or code"));
         }
       }
       WHEN("batch register is called with kVariant errorPolicyTag") {
@@ -1003,105 +874,6 @@ SCENARIO("Registering entities") {
           CHECK(error1 == expectedError1);
 
           CHECK(std::get<openassetio::EntityReference>(actualVec[2]) == expectedValue2);
-        }
-      }
-    }
-  }
-}
-
-TEMPLATE_TEST_CASE(
-    "BatchElementError conversion to exceptions when registering", "",
-    (BatchElementErrorMapping<openassetio::errors::UnknownBatchElementException,
-                              ErrorCode::kUnknown>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidEntityReferenceBatchElementException,
-                              ErrorCode::kInvalidEntityReference>),
-    (BatchElementErrorMapping<openassetio::errors::MalformedEntityReferenceBatchElementException,
-                              ErrorCode::kMalformedEntityReference>),
-    (BatchElementErrorMapping<openassetio::errors::EntityAccessErrorBatchElementException,
-                              ErrorCode::kEntityAccessError>),
-    (BatchElementErrorMapping<openassetio::errors::EntityResolutionErrorBatchElementException,
-                              ErrorCode::kEntityResolutionError>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidPreflightHintBatchElementException,
-                              ErrorCode::kInvalidPreflightHint>),
-    (BatchElementErrorMapping<openassetio::errors::InvalidTraitSetBatchElementException,
-                              ErrorCode::kInvalidTraitSet>)) {
-  namespace hostApi = openassetio::hostApi;
-  using trompeloeil::_;
-
-  using ExpectedExceptionType = typename TestType::ExceptionType;
-  static constexpr ErrorCode kExpectedErrorCode = TestType::kErrorCode;
-
-  GIVEN("a configured Manager instance") {
-    const openassetio::trait::TraitSet traits = {"fakeTrait", "secondFakeTrait"};
-    const openassetio::ManagerFixture fixture;
-    const auto& manager = fixture.manager;
-    auto& mockManagerInterface = fixture.mockManagerInterface;
-    const auto& context = fixture.context;
-    const auto& hostSession = fixture.hostSession;
-    const auto publishingAccess = openassetio::access::PublishingAccess::kWrite;
-
-    AND_GIVEN(
-        "manager plugin will encounter an entity-specific error when next registering a "
-        "reference") {
-      const openassetio::EntityReference ref = openassetio::EntityReference{"testReference"};
-      const openassetio::EntityReferences refs = {ref};
-
-      auto singleTraitsData = openassetio::TraitsData::make(traits);
-      const openassetio::trait::TraitsDatas singleTraitsDatas = {singleTraitsData};
-
-      const openassetio::errors::BatchElementError expectedError{kExpectedErrorCode,
-                                                                 "Some error message"};
-
-      // With error callback side effect
-      REQUIRE_CALL(mockManagerInterface, register_(refs, singleTraitsDatas, publishingAccess,
-                                                   context, hostSession, _, _))
-          .LR_SIDE_EFFECT(_7(123, expectedError));
-
-      WHEN("register is called with kException errorPolicyTag") {
-        THEN("an exception is thrown") {
-          try {
-            manager->register_(ref, singleTraitsData, publishingAccess, context,
-                               hostApi::Manager::BatchElementErrorPolicyTag::kException);
-            FAIL_CHECK("Exception not thrown");
-          } catch (const ExpectedExceptionType& exc) {
-            CHECK(exc.what() == expectedError.message);
-            CHECK(exc.error == expectedError);
-            CHECK(exc.index == 123);
-          }
-        }
-      }
-    }
-
-    AND_GIVEN(
-        "manager plugin will encounter entity-specific errors when next registering multiple "
-        "references") {
-      const openassetio::trait::TraitsDatas threeTraitsDatas = {
-          openassetio::TraitsData::make(traits), openassetio::TraitsData::make(traits),
-          openassetio::TraitsData::make(traits)};
-
-      const openassetio::EntityReferences refs = {openassetio::EntityReference{"ref1"},
-                                                  openassetio::EntityReference{"ref2"},
-                                                  openassetio::EntityReference{"ref3"}};
-
-      const openassetio::errors::BatchElementError expectedError{kExpectedErrorCode, "Some error"};
-
-      // With error callback side effect
-      REQUIRE_CALL(mockManagerInterface,
-                   register_(refs, threeTraitsDatas, publishingAccess, context, hostSession, _, _))
-          .LR_SIDE_EFFECT(_7(123, expectedError))
-          .LR_SIDE_EFFECT(FAIL_CHECK("Exception should have short-circuited this"));
-
-      WHEN("register is called with kException errorPolicyTag") {
-        THEN("an exception is thrown") {
-          try {
-            manager->register_(refs, threeTraitsDatas, publishingAccess, context,
-                               hostApi::Manager::BatchElementErrorPolicyTag::kException);
-            FAIL_CHECK("Exception not thrown");
-          } catch (const ExpectedExceptionType& exc) {
-            CHECK(exc.what() == expectedError.message);
-            CHECK(exc.error == expectedError);
-            CHECK(exc.index == 123);
-          }
         }
       }
     }