+++ src/openassetio-python/cmodule/src/errors/exceptionsBinding.cpp
@@ -77,7 +77,7 @@ void registerNonBatchExceptions(const py::module &mod) {
   });
 }
 
-void registerBatchElementExceptions(const py::module &mod) {
+void registerBatchElementException(const py::module &mod) {
   // Pybind has very limited support for custom exception types. This is
   // a well-known tricky issue and is apparently not on the roadmap to
   // fix. The only direct support is for an exception type that takes a
@@ -87,21 +87,18 @@ void registerBatchElementExceptions(const py::module &mod) {
   // A way forward is provided by the sketch in:
   // https://github.com/pybind/pybind11/issues/1281#issuecomment-1375950333
   //
-  // We execute a Python string literal to create our base exception
+  // We execute a Python string literal to create our exception
   // type. The `globals` and `locals` dict parameters dictate the scope
   // of execution, so we use this to ensure the definition is scoped to
   // our `_openassetio` module.
-  //
-  // We can then retrieve this base exception type as a pybind object
-  // and use it as the base class for pybind's limited exception
-  // registration API.
 
   py::exec(R"pybind(
 class BatchElementException(RuntimeError):
-    def __init__(self, index: int, error):
+    def __init__(self, index: int, error, message: str):
         self.index = index
         self.error = error
-        super().__init__(error.message))pybind",
+        self.message = message
+        super().__init__(message))pybind",
            mod.attr("__dict__"), mod.attr("__dict__"));
 
   // Retrieve a handle the the exception type just created by executing
@@ -109,30 +106,6 @@ class BatchElementException(RuntimeError):
   const py::object pyBatchElementException = mod.attr("BatchElementException");
 
   using openassetio::errors::BatchElementException;
-  using openassetio::errors::EntityAccessErrorBatchElementException;
-  using openassetio::errors::EntityResolutionErrorBatchElementException;
-  using openassetio::errors::InvalidEntityReferenceBatchElementException;
-  using openassetio::errors::InvalidPreflightHintBatchElementException;
-  using openassetio::errors::InvalidTraitSetBatchElementException;
-  using openassetio::errors::MalformedEntityReferenceBatchElementException;
-  using openassetio::errors::UnknownBatchElementException;
-
-  // Register a new exception type using `BatchElementException` base
-  // class created above. Note that this is not sufficient to cause C++
-  // exceptions to be translated. See `register_exception_translator`
-  // below.
-  const auto registerPyException = [&mod, &pyBatchElementException](const char *pyTypeName) {
-    py::exception<void /* unused */>{mod, pyTypeName, pyBatchElementException};
-  };
-
-  // Register each of our BatchElement exception types.
-  registerPyException("UnknownBatchElementException");
-  registerPyException("InvalidEntityReferenceBatchElementException");
-  registerPyException("MalformedEntityReferenceBatchElementException");
-  registerPyException("EntityAccessErrorBatchElementException");
-  registerPyException("EntityResolutionErrorBatchElementException");
-  registerPyException("InvalidPreflightHintBatchElementException");
-  registerPyException("InvalidTraitSetBatchElementException");
 
   // Register a function that will translate our C++ exceptions to the
   // appropriate Python exception type.
@@ -149,7 +122,7 @@ class BatchElementException(RuntimeError):
     // the currently active Python exception in this thread.
     const auto setPyException = [&pyModule](const char *pyTypeName, const auto &exc) {
       const py::object pyClass = pyModule.attr(pyTypeName);
-      const py::object pyInstance = pyClass(exc.index, exc.error);
+      const py::object pyInstance = pyClass(exc.index, exc.error, exc.what());
       PyErr_SetObject(pyClass.ptr(), pyInstance.ptr());
     };
 
@@ -158,20 +131,6 @@ class BatchElementException(RuntimeError):
     // exception in this thread.
     try {
       std::rethrow_exception(std::move(pexc));
-    } catch (const UnknownBatchElementException &exc) {
-      setPyException("UnknownBatchElementException", exc);
-    } catch (const InvalidEntityReferenceBatchElementException &exc) {
-      setPyException("InvalidEntityReferenceBatchElementException", exc);
-    } catch (const MalformedEntityReferenceBatchElementException &exc) {
-      setPyException("MalformedEntityReferenceBatchElementException", exc);
-    } catch (const EntityAccessErrorBatchElementException &exc) {
-      setPyException("EntityAccessErrorBatchElementException", exc);
-    } catch (const EntityResolutionErrorBatchElementException &exc) {
-      setPyException("EntityResolutionErrorBatchElementException", exc);
-    } catch (const InvalidPreflightHintBatchElementException &exc) {
-      setPyException("InvalidPreflightHintBatchElementException", exc);
-    } catch (const InvalidTraitSetBatchElementException &exc) {
-      setPyException("InvalidTraitSetBatchElementException", exc);
     } catch (const BatchElementException &exc) {
       setPyException("BatchElementException", exc);
     }
@@ -183,5 +142,5 @@ void registerExceptions(const py::module &mod) {
   registerNonBatchExceptions(mod);
   // Make sure the non batch exception is registered first, as that
   // includes the shared base exception.
-  registerBatchElementExceptions(mod);
+  registerBatchElementException(mod);
 }