+++ src/openassetio-python/package/openassetio/errors.py
@@ -27,14 +27,3 @@ NotImplementedException = _openassetio.errors.NotImplementedException
 UnhandledException = _openassetio.errors.UnhandledException
 BatchElementError = _openassetio.errors.BatchElementError
 BatchElementException = _openassetio.errors.BatchElementException
-UnknownBatchElementException = _openassetio.errors.UnknownBatchElementException
-InvalidTraitSetBatchElementException = _openassetio.errors.InvalidTraitSetBatchElementException
-EntityAccessErrorBatchElementException = _openassetio.errors.EntityAccessErrorBatchElementException
-# Black will format these long ones as a multiline scope, which doxygen
-# mistakes as a tuple.
-# fmt: off
-InvalidEntityReferenceBatchElementException = _openassetio.errors.InvalidEntityReferenceBatchElementException
-MalformedEntityReferenceBatchElementException = _openassetio.errors.MalformedEntityReferenceBatchElementException
-EntityResolutionErrorBatchElementException = _openassetio.errors.EntityResolutionErrorBatchElementException
-InvalidPreflightHintBatchElementException = _openassetio.errors.InvalidPreflightHintBatchElementException
-# fmt: on