+++ RELEASE_NOTES.md
@@ -10,6 +10,11 @@ v1.0.0-alpha.xx
   share a common base.
   [#1071](https://github.com/OpenAssetIO/OpenAssetIO/issues/1071)
 
+- Exceptional conveniences now provide additional information such
+  as entityReference, index, access, and error type in the message
+  of th exception.
+  [#1073](https://github.com/OpenAssetIO/OpenAssetIO/issues/1073)
+
 ### Breaking changes
 
 - Attempting to retrieve a trait property with
@@ -32,6 +37,11 @@ v1.0.0-alpha.xx
 
 - Removed python `exceptions` module in favour of new `errors` module.
   [#1071](https://github.com/OpenAssetIO/OpenAssetIO/issues/1071)
+
+- Removed all subtypes of `BatchElementException`, in favour of a single
+  exception type for exceptional type for batch errors in exceptional
+  conveniences.
+  [#1073](https://github.com/OpenAssetIO/OpenAssetIO/issues/1073)
 
 v1.0.0-alpha.14
 ---------------